+++ packages/react/src/components/table/table.stories.ts
@@ -1,0 +1,16 @@
+import { storiesOfTable } from '@dso-toolkit/sources';
+import { storiesOf } from '@storybook/react';
+
+import readme from './readme.md';
+import { tableTemplate } from './table.template';
+
+storiesOfTable(
+  {
+    module,
+    storiesOf,
+    readme
+  },
+  {
+    tableTemplate
+  }
+);