+++ packages/react/src/components/table/table.template.tsx
@@ -1,0 +1,12 @@
+import { Table } from '@dso-toolkit/sources';
+import * as React from 'react';
+
+import { DsoTable } from '../..';
+
+export function tableTemplate({ responsive }: Table) {
+  return (
+    <DsoTable
+      responsive={responsive}
+    />
+  );
+}