+++ packages/sources/src/components/table/table.args.ts
@@ -1,0 +1,18 @@
+import { ArgTypes } from '../../storybook';
+import { Table } from './table.models';
+
+export interface TableArgs {
+  responsive: boolean;
+}
+
+export const tableArgTypes: ArgTypes<TableArgs> = {
+  responsive: {
+    control: {
+      type: 'boolean'
+    }
+  },
+};
+
+export function tableArgsMapper(a: TableArgs): Required<Table> {
+  return { ...a };
+}